whatigot
setup currency mode toggle button and currency switch slot... switched listeners so players can move items in their own inventory. currently if they drag to the switch currency spot it will make a ghost object in that slot, but updating the config with item info is not implemented... prob not 1.4.2?